Extract Overall Page Metrics utility functions to separate files.
