+++ assets/js/modules/analytics/components/dashboard/DashboardOverallPageMetricsWidget/index.js
@@ -19,223 +19,24 @@
 /**
  * WordPress dependencies
  */
-import { __, _x, sprintf, _n } from '@wordpress/i18n';
-import { isURL } from '@wordpress/url';
+import { _x, sprintf, _n } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import {
-	DATE_RANGE_OFFSET,
-	MODULES_ANALYTICS,
-} from '../../../datastore/constants';
+import { MODULES_ANALYTICS } from '../../../datastore/constants';
 import { CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
-import { CORE_SITE } from '../../../../../googlesitekit/datastore/site/constants';
 import { Grid, Row, Cell } from '../../../../../material-components/layout';
 import PreviewBlock from '../../../../../components/PreviewBlock';
 import DataBlock from '../../../../../components/DataBlock';
 import Sparkline from '../../../../../components/Sparkline';
 import SourceLink from '../../../../../components/SourceLink';
 import whenActive from '../../../../../util/when-active';
-import { generateDateRangeArgs } from '../../../util/report-date-range-args';
-import { calculateChange, getURLPath } from '../../../../../util';
-import parseDimensionStringToDate from '../../../util/parseDimensionStringToDate';
 import WidgetHeaderTitle from '../../../../../googlesitekit/widgets/components/WidgetHeaderTitle';
-import useViewOnly from '../../../../../hooks/useViewOnly';
+import { useOverallPageMetricsReport } from './hooks/useOverallPageMetricsReport';
+import { calculateOverallPageMetricsData } from './utils';
 const { useSelect, useInViewSelect } = Data;
-
-/**
- * Fetches Analytics report data and state for the Overall Page Metrics widget.
- *
- * @since 1.45.0
- *
- * @typedef {Object} OverallPageMetricsReport
- * @property {Array.<Object>|undefined} report     - Analytics report data if exists, otherwise undefined.
- * @property {string}                   serviceURL - Link to relevant Google Analytics page for the report.
- * @property {boolean}                  isLoading  - Loading status for report.
- * @property {(Object|undefined)}       error      - Error object if exists, otherwise undefined.
- * @return {OverallPageMetricsReport} Analytics report data and state.
- */
-function useOverallPageMetricsReport() {
-	const viewOnlyDashboard = useViewOnly();
-
-	const dates = useSelect( ( select ) =>
-		select( CORE_USER ).getDateRangeDates( {
-			offsetDays: DATE_RANGE_OFFSET,
-			compare: true,
-		} )
-	);
-
-	const url = useSelect( ( select ) =>
-		select( CORE_SITE ).getCurrentEntityURL()
-	);
-
-	const args = {
-		...dates,
-		dimensions: [ 'ga:date' ],
-		metrics: [
-			{
-				expression: 'ga:pageviews',
-				alias: 'Pageviews',
-			},
-			{
-				expression: 'ga:uniquePageviews',
-				alias: 'Unique Pageviews',
-			},
-			{
-				expression: 'ga:bounceRate',
-				alias: 'Bounce Rate',
-			},
-			{
-				expression: 'ga:avgSessionDuration',
-				alias: 'Session Duration',
-			},
-		],
-		url,
-	};
-
-	const reportArgs = generateDateRangeArgs( dates );
-
-	if ( isURL( url ) ) {
-		reportArgs[ 'explorer-table.plotKeys' ] = '[]';
-		reportArgs[ '_r.drilldown' ] = `analytics.pagePath:${ getURLPath(
-			url
-		) }`;
-	}
-
-	const isLoading = useSelect(
-		( select ) =>
-			! select( MODULES_ANALYTICS ).hasFinishedResolution( 'getReport', [
-				args,
-			] )
-	);
-
-	const error = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS ).getErrorForSelector( 'getReport', [ args ] )
-	);
-
-	const serviceURL = useSelect( ( select ) => {
-		if ( viewOnlyDashboard ) {
-			return null;
-		}
-
-		return select( MODULES_ANALYTICS ).getServiceReportURL(
-			'visitors-overview',
-			reportArgs
-		);
-	} );
-
-	const report = useInViewSelect( ( select ) =>
-		select( MODULES_ANALYTICS ).getReport( args )
-	);
-
-	return {
-		report,
-		serviceURL,
-		isLoading,
-		error,
-	};
-}
-
-/**
- * Parse Analytics report into data suitable for rendering the data blocks in the Overall Page Metrics widget.
- *
- * @typedef {Object} OverallPageMetricsData
- * @property {string}         metric          - Google Analytics metric identifier.
- * @property {string}         title           - Translated metric title.
- * @property {Array.<Object>} sparkLineData   - Data for rendering the sparkline.
- * @property {string}         [datapointUnit] - Optional datapoint unit, e.g. '%', 's'.
- * @property {number}         total           - Total count for the metric.
- * @property {number}         change          - Monthly change for the metric.
- *
- * @since 1.45.0
- *
- * @param {Object} report Analytics report data.
- * @return {Array.<OverallPageMetricsData>} Array of data for rendering the data blocks in the Overall Page Metrics widget.
- */
-
-function calculateOverallPageMetricsData( report ) {
-	const metricsData = [
-		{
-			metric: 'ga:pageviews',
-			title: __( 'Pageviews', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Pageviews' },
-				],
-			],
-			total: 0,
-			change: 0,
-		},
-		{
-			metric: 'ga:uniquePageviews',
-			title: __( 'Unique Pageviews', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Unique Pageviews' },
-				],
-			],
-			total: 0,
-			change: 0,
-		},
-		{
-			metric: 'ga:bounceRate',
-			title: __( 'Bounce Rate', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Bounce Rate' },
-				],
-			],
-			datapointUnit: '%',
-			datapointDivider: 100,
-			total: 0,
-			change: 0,
-		},
-		{
-			metric: 'ga:avgSessionDuration',
-			title: __( 'Session Duration', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Session Duration' },
-				],
-			],
-			datapointUnit: 's',
-			total: 0,
-			change: 0,
-		},
-	];
-
-	const { totals = [], rows = [] } = report?.[ 0 ]?.data || {};
-
-	const lastMonth = totals[ 0 ]?.values || [];
-	const previousMonth = totals[ 1 ]?.values || [];
-
-	return metricsData.map(
-		( { datapointDivider = 1, ...metricData }, index ) => {
-			// We only want half the date range, having a comparison date range in the query doubles the range.
-			for ( let i = Math.ceil( rows.length / 2 ); i < rows.length; i++ ) {
-				const { values } = rows[ i ].metrics[ 0 ];
-				const dateString = rows[ i ].dimensions[ 0 ];
-				const date = parseDimensionStringToDate( dateString );
-
-				metricData.sparkLineData.push( [ date, values[ index ] ] );
-			}
-
-			metricData.total = ( lastMonth[ index ] || 0 ) / datapointDivider;
-			metricData.change = calculateChange(
-				previousMonth[ index ] || 0,
-				lastMonth[ index ] || 0
-			);
-
-			return metricData;
-		}
-	);
-}
 
 function DashboardOverallPageMetricsWidget( { Widget, WidgetReportError } ) {
 	const isGatheringData = useInViewSelect( ( select ) =>
