+++ assets/js/modules/analytics/components/dashboard/DashboardOverallPageMetricsWidgetGA4/index.js
@@ -19,19 +19,13 @@
 /**
  * WordPress dependencies
  */
-import { __, _x, sprintf, _n } from '@wordpress/i18n';
-import { isURL } from '@wordpress/url';
+import { _x, sprintf, _n } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import {
-	DATE_RANGE_OFFSET,
-	MODULES_ANALYTICS,
-} from '../../../datastore/constants';
 import { CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
-import { CORE_SITE } from '../../../../../googlesitekit/datastore/site/constants';
 import { MODULES_ANALYTICS_4 } from '../../../../analytics-4/datastore/constants';
 import { Grid, Row, Cell } from '../../../../../material-components/layout';
 import PreviewBlock from '../../../../../components/PreviewBlock';
@@ -39,231 +33,10 @@ import DataBlock from '../../../../../components/DataBlock';
 import Sparkline from '../../../../../components/Sparkline';
 import SourceLink from '../../../../../components/SourceLink';
 import whenActive from '../../../../../util/when-active';
-import { generateDateRangeArgs } from '../../../util/report-date-range-args';
-import { calculateChange, getURLPath, stringToDate } from '../../../../../util';
-import parseDimensionStringToDate from '../../../util/parseDimensionStringToDate';
 import WidgetHeaderTitle from '../../../../../googlesitekit/widgets/components/WidgetHeaderTitle';
-import useViewOnly from '../../../../../hooks/useViewOnly';
+import { useOverallPageMetricsReport } from './hooks/useOverallPageMetricsReport';
+import { calculateOverallPageMetricsData } from './utils';
 const { useSelect, useInViewSelect } = Data;
-
-/**
- * Fetches Analytics 4 report data and state for the Overall Page Metrics widget.
- *
- * @since n.e.x.t
- *
- * @typedef {Object} OverallPageMetricsReport
- * @property {Object|undefined}   report     - Analytics report data if exists, otherwise undefined.
- * @property {Object}             dates      - Object containing the report date ranges.
- * @property {string}             serviceURL - Link to relevant Google Analytics page for the report.
- * @property {boolean}            isLoading  - Loading status for report.
- * @property {(Object|undefined)} error      - Error object if exists, otherwise undefined.
- * @return {OverallPageMetricsReport} Analytics report data and state.
- */
-function useOverallPageMetricsReport() {
-	const viewOnlyDashboard = useViewOnly();
-
-	const dates = useSelect( ( select ) =>
-		select( CORE_USER ).getDateRangeDates( {
-			offsetDays: DATE_RANGE_OFFSET,
-			compare: true,
-		} )
-	);
-
-	const url = useSelect( ( select ) =>
-		select( CORE_SITE ).getCurrentEntityURL()
-	);
-
-	const args = {
-		...dates,
-		dimensions: [ 'date' ],
-		metrics: [
-			{
-				name: 'screenPageViews',
-			},
-			{
-				name: 'sessions',
-			},
-			{
-				name: 'engagedSessions',
-			},
-			{
-				name: 'averageSessionDuration',
-			},
-		],
-		orderby: [
-			{
-				dimension: {
-					dimensionName: 'date',
-				},
-			},
-		],
-		url,
-	};
-
-	const reportArgs = generateDateRangeArgs( dates );
-
-	if ( isURL( url ) ) {
-		reportArgs[ 'explorer-table.plotKeys' ] = '[]';
-		reportArgs[ '_r.drilldown' ] = `analytics.pagePath:${ getURLPath(
-			url
-		) }`;
-	}
-
-	const isLoading = useSelect(
-		( select ) =>
-			! select( MODULES_ANALYTICS_4 ).hasFinishedResolution(
-				'getReport',
-				[ args ]
-			)
-	);
-
-	const error = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS_4 ).getErrorForSelector( 'getReport', [
-			args,
-		] )
-	);
-
-	const serviceURL = useSelect( ( select ) => {
-		if ( viewOnlyDashboard ) {
-			return null;
-		}
-
-		return select( MODULES_ANALYTICS ).getServiceReportURL(
-			'visitors-overview',
-			reportArgs
-		);
-	} );
-
-	const report = useInViewSelect( ( select ) =>
-		select( MODULES_ANALYTICS_4 ).getReport( args )
-	);
-
-	return {
-		report,
-		dates,
-		serviceURL,
-		isLoading,
-		error,
-	};
-}
-
-/**
- * Parse Analytics 4 report into data suitable for rendering the data blocks in the Overall Page Metrics widget.
- *
- * @typedef {Object} OverallPageMetricsData
- * @property {string}         metric          - Google Analytics metric identifier.
- * @property {string}         title           - Translated metric title.
- * @property {Array.<Object>} sparkLineData   - Data for rendering the sparkline.
- * @property {string}         [datapointUnit] - Optional datapoint unit, e.g. '%', 's'.
- * @property {number}         total           - Total count for the metric.
- * @property {number}         change          - Monthly change for the metric.
- *
- * @since n.e.x.t
- *
- * @param {Object} report    Analytics report data.
- * @param {string} startDate Start date for the report.
- * @return {Array.<OverallPageMetricsData>} Array of data for rendering the data blocks in the Overall Page Metrics widget.
- */
-
-function calculateOverallPageMetricsData( report, startDate ) {
-	const metricsData = [
-		{
-			metric: 'screenPageViews',
-			title: __( 'Pageviews', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Pageviews' },
-				],
-			],
-			total: 0,
-			change: 0,
-		},
-		{
-			metric: 'sessions',
-			title: __( 'Sessions', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Sessions' },
-				],
-			],
-			total: 0,
-			change: 0,
-		},
-		{
-			metric: 'engagedSessions',
-			title: __( 'Engaged Sessions', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Engaged Sessions' },
-				],
-			],
-			datapointUnit: '%',
-			datapointDivider: 100,
-			total: 0,
-			change: 0,
-		},
-		{
-			metric: 'averageSessionDuration',
-			title: __( 'Session Duration', 'google-site-kit' ),
-			sparkLineData: [
-				[
-					{ type: 'date', label: 'Day' },
-					{ type: 'number', label: 'Session Duration' },
-				],
-			],
-			datapointUnit: 's',
-			total: 0,
-			change: 0,
-		},
-	];
-
-	const { totals = [], rows = [] } = report || {};
-
-	const lastMonth = totals[ 0 ]?.metricValues || [];
-	const previousMonth = totals[ 1 ]?.metricValues || [];
-
-	const startDateTime = stringToDate( startDate ).getTime();
-	const currentDateRangeRows = rows.filter( ( { dimensionValues } ) => {
-		if ( dimensionValues[ 1 ].value !== 'date_range_0' ) {
-			return false;
-		}
-
-		// We only want half of the date range, as having a comparison date range in the query doubles the range.
-		// In order to achieve this, we filter out entries before the start date (the comparison start date will be earlier).
-		const rowDate = parseDimensionStringToDate(
-			dimensionValues[ 0 ].value
-		);
-		return rowDate.getTime() >= startDateTime;
-	} );
-
-	return metricsData.map(
-		( { datapointDivider = 1, ...metricData }, index ) => {
-			currentDateRangeRows.forEach(
-				( { dimensionValues, metricValues } ) => {
-					const dateString = dimensionValues[ 0 ].value;
-					const date = parseDimensionStringToDate( dateString );
-
-					metricData.sparkLineData.push( [
-						date,
-						metricValues[ index ].value,
-					] );
-				}
-			);
-
-			metricData.total =
-				( lastMonth[ index ]?.value || 0 ) / datapointDivider;
-			metricData.change = calculateChange(
-				previousMonth[ index ]?.value || 0,
-				lastMonth[ index ]?.value || 0
-			);
-
-			return metricData;
-		}
-	);
-}
 
 function DashboardOverallPageMetricsWidgetGA4( { Widget, WidgetReportError } ) {
 	const isGatheringData = useInViewSelect( ( select ) =>
